Move console module to binary
