+++ src/bin/console/highlighter.rs
