+++ src/bin/console/validator.rs
