+++ src/bin/diatom.rs
@@ -1,7 +1,12 @@
-use diatom::{Console, Interpreter};
+use diatom::Interpreter;
 use std::{fs, io, path::PathBuf};
 
 use clap::Parser;
+
+#[cfg(feature = "console")]
+mod console;
+#[cfg(feature = "console")]
+pub use console::Console;
 
 #[derive(Parser)]
 #[command(name = "Diatom Interpreter")]
@@ -28,9 +33,14 @@ fn main() {
     let args = Args::parse();
 
     match (&args.path, args.inspect) {
+        #[cfg(feature = "console")]
         (None, inspect) => {
             let mut console = Console::new(!args.no_color);
             console.run(inspect);
+        }
+        #[cfg(not(feature = "console"))]
+        (None, _) => {
+            eprintln!("Error: Diatom is not compiled with console enabled")
         }
         (Some(path), false) => {
             let code = fs::read_to_string(path).expect("Error: File can not be read!");
